+++ MMCore/MMCore.h
@@ -394,12 +394,12 @@ public:
    double getExposure() MMCORE_LEGACY_THROW(CMMError);
    double getExposure(const char* label) MMCORE_LEGACY_THROW(CMMError);
 
-   int getBinning() MMCORE_LEGACY_THROW(CMMError);
-   int getBinning(const char* label) MMCORE_LEGACY_THROW(CMMError);
-   std::vector<int> getAllowedBinningValues() MMCORE_LEGACY_THROW(CMMError);
-   std::vector<int> getAllowedBinningValues(const char* label) MMCORE_LEGACY_THROW(CMMError);
-   void setBinning(int binning) MMCORE_LEGACY_THROW(CMMError);
-   void setBinning(const char* label, int binning) MMCORE_LEGACY_THROW(CMMError);
+   long getBinning() MMCORE_LEGACY_THROW(CMMError);
+   long getBinning(const char* label) MMCORE_LEGACY_THROW(CMMError);
+   std::vector<long> getAllowedBinningValues() MMCORE_LEGACY_THROW(CMMError);
+   std::vector<long> getAllowedBinningValues(const char* label) MMCORE_LEGACY_THROW(CMMError);
+   void setBinning(long binning) MMCORE_LEGACY_THROW(CMMError);
+   void setBinning(const char* label, long binning) MMCORE_LEGACY_THROW(CMMError);
 
    void snapImage() MMCORE_LEGACY_THROW(CMMError);
    void* getImage() MMCORE_LEGACY_THROW(CMMError);
